requirements.txt
lxml==5.3.0
packaging==24.2
pdf2image==1.17.0
pillow==11.0.0
PyMuPDF==1.24.13
pytesseract==0.3.13
python-docx==1.1.2
python-pptx==1.0.2
typing_extensions==4.12.2
XlsxWriter==3.2.0
